Add class MetadataBlock.
Adds class MetadataBlock, which is a storage container for the
(meta)data related to a file.
The idea is that the fields can be accessed directly, and having it be a
class allows for future expansion if need be. Some checking might need
to be done before modifying a field, etc. Aside from that, the class is
meant to perform like a dictionary, I.E. basic key-value storage.