Fix datetime extraction from text contents.
Fixes datetime extraction from text contents by pulling in code from
'get_datetime_from_text' in 'dateandtime.py' that was moved there to
make the text extraction "general-purpose" and able to handle any kind
of textual content. The idea is sound but was not implemented well or
even at all.