Store relevant file mime types in the analyzers.
Moves the lookup for which mime type are relevant to which analyzers to
the analyzers themselves.
Adds a new field 'applies_to_mime' to the abstract baseclass
'AbstractAnalyzer' and all subclasses.  The field can be a string, a
list of strings or None, which means the analyzer should be used
regardless of file mimetype.