Fix usage information message bug.
Fixes bug that caused the message 'Add "--help" to display usage
information.' was displayed even when '--help' was specified.

Caused by incorrect number in argument length test. Fixed by checking
that the list is non-zero, seems to work the same as explicitly testing
if the argument list length is non-zero.