Attempt at fixing launch script path resolve.
Partial fix for resolving absolute paths in the launcher script.

The MacOS (BSD) version of 'readlink' does not have the '-f' option.

This change adds a test of which version of 'readlink' is available and
a partial fix for MacOS. Still does not handle multiple layers, or
nested symlinks properly.