Add TODO-note on weighting date/time results.
Adds TODO-notes on including the number of tries needed in order to find
valid date/time data from some input. This number would be included with
the returned date/time data, possibly as a tuple.

The number would after normalization function as a weight that would be
use to rank the result and calculate the probability that the data is
actually correct.