Store raw stdout to the log file.
On second thought..

The Python "pytest" test results can be exported to HTML from PyCharm.

If I store the raw output from the bash tests to the log file, I can use
the program 'aha' to convert the captured stdout to HTML, preserving
formatting and coloring.

Which is more consistent and should hopefully be easier to consolidate.