Add main test runner 'tests/run_all_tests.sh'.
Adds a new file 'run_all_tests.sh' which is the main test runner script
for running all tests.

This script calls other "test suites" which perform the actual tests.
All files in the 'tests' directory matching "test_*.sh" are considered
test suites to be executed.

This script keeps track of the total execution time and final conversion
of the test output log to HTML.