Fix bug where incorrect filename is printed.
Fixes a bug where the testing script filename is printed
instead of the file 'common_utils.sh' that contains the
named variable "AUTONAMEOW_WIKI_ROOT_DIR".

This was caused by referring to the variable using the variable "SELF"
which is changed by the sourcing test scripts that prints the message.