Fix bug where getting script path fails sourced.
Fixes a bug where getting the full path to he executing script failed
when the script was sourced.

Getting the script name through "$0" does not work when the script is
sourced. Adds convoluted workaround that seems to be able to resolve the
full absolute path to the script no matter how it is used.