Major restructuring of module hierarchy.
Adds a major restructuring of the module hierarchy.

All analyzers are moved to a new module 'analyzers' located at
the same "depth" as the main 'core' module.

This is motivated by reflecting the intended (coming) design where
there is a strict separation between the main program core and
auxiliary modules that perform various tasks like data extraction
and interaction with other systems.

The code base will be split into at least two main parts;
the main 'core' and "the rest", which is the 'analyzers' for now.