Remove queueing default analyzers.
Removes adding default analyzers to the run queue.

The analyzers 'FilesystemAnalyzer' and 'FilenameAnalyzer' was added to
the analysis run queue as part of the 'Analysis' class initialization.

This is now removed, these analyzers are added for all files at a later
stage because their 'applies_to_mime' is set to 'MIME_ALL'.