Remove 'FileRule' class '__eq__' comparison.
Removing a file rule from a list of file rules used this method,
which caused errors as any rule with the same score and weight
as the rule to remove was at risk for erroneous removal.

Removing this makes list removal fall back to defaults.

Sorting still works, which is the only reson the comparison
methods were added in the first place.