Remove duplicate tests for file access.
Removes testing for read permission to the configuration file.

When the path to the file to read is specified with the option
'--config-path', the argument parser performs this test among
other tests. So this case is covered.

When the path comes from default configuration paths or
sources given by other potential callers, the read might fail
if not caught by 'IOError'.

This should be investigated further.