Refactor gathering results in 'get_publisher'.
Adds questionable refactoring of data collection in the
'get_publisher' method of the 'PdfAnalyzer' class.

These "getter"-methods need to be redesigned or possibly removed.
But I am not at all sure that the changes added in this commit are
the way to go or even headed in roughly the right direction..