Handle exception in the 'dump' function.
Adds handling exception thrown by 'yaml' in the 'dump' function.

The cause of this error is unclear at this point.
Some input data causes the error.  I investigated this by stepping
through the 'yaml' code with input data known to cause the error.

I did not manage to pinpoint which data caused the error, neither
if the problem code is part of the autonameow core or part of the
results returned by 'exiftool'.