Implement (partial) data extraction.
Adds a partial implementation of data extraction in 'Extraction'.

Work in progress changes to implement the data extraction handler
class 'Extraction' with a queue of "extractors" suitable for a
given file.

A queue is populated with extractor classes whose class attribute
'handles_mime_type' contains the MIME type of the given file.

Instances of the enqueued extractor classes are instantiated
and subsequently executed. Results are stored in 'self.data'
which is an instance of the 'ExtractedData' class.

Each result is stored under a label defined by the class attribute
'data_query_string' of each extractor instance.