Improve querying for analysis results data.
Improves how analysis data is stored and retrieved.

Data received through callbacks in the 'Analysis' class is
pre-processed before being stored in the 'AnalysisResults' class.
The "pre-processing" is simply to "flatten" any incoming dictionary
in the 'Analysis' class 'collect_results' method.

    Incoming arguments:
    LABEL: 'metadata.exiftool'     DATA: {'a': 'b', 'c': 'd'}

    Would be "flattened" to:
    LABEL: 'metadata.exiftool.a'   DATA: 'b'
    LABEL: 'metadata.exiftool.c'   DATA: 'd'

This allows fixing temporary fixes in the 'query' method, that is
used to retrieve data from the 'AnalysisResults' class.