Modify testing extractor class suitability.
Modifies how extractor classes are deemed suitable for a file.

Old method evaluated the given file MIME type against the MIME type
blobs in each of the classes attribute 'handles_mime_types'.

This change adds testing suitability by a new class method 'can_handle',
defined in the 'Extractor' base class. This method takes a 'FileObject'
and returns a boolean.

The internal implementation is the same, but this change opens up for
future overriding of the 'can_handle' method in each of the extractor
classes if they require additional testing aside from just the file
MIME type.