Modify testing analyzer class suitability.
Modifies how analyzer classes are deemed suitable for a given file.

The old method evaluated the MIME type of the given file against the
MIME type blobs in each of the analyzer class attribute
'handles_mime_types'.

This change adds a new class method 'can_handle' that performs this
test.  The internal implementation is the same, with an advantage:

This method is defined in the 'Extractor' base class.
Any future extractor classes that requires additional testing, aside
from just the file MIME type, can override the 'can_handle' method.