Disable passing 'pdftotext' results to 'unidecode'.
Disable passing 'pdftotext' results to 'unidecode' in 'sanitize_text'.

This method strips all umlauts when running on my desktop.
Behaviour is completely different when running on my MacbookPro.
This has to do with not handling character encoding properly.

This is a critical item on the TODO-list which should be fixed ASAP.