Add (WIP) rework of evaluating file rule conditions.
Adds incomplete rework of how file rules are evaluated.

Adds modifications to handle the 'RuleCondition' classes instead of the
previously used dicts. The advantage of this method is that the
'RuleCondition' instances can store a reference to a suitable parser
that can be called to perform the evaluation.

This will replace the old mess of "if-else" statements.