Add (WIP) rework to solve mutable 'FileRule' state.
Adds a (WIP) fix for mutable 'FileRule' unwantingly shared state.

The 'FileRule' class score is shared between runs of different files.
This adds a in progress fix that moves the score to a 'Counter' that is
keyed by the 'FileRule' class instances, with the keys storing the
number of conditions that were met for the current file.

The 'RuleMatcher' instance is recreated for each file, so the state will
be cleared and the problem will be solved in a good enough way..