Add simpler solution to the mutable state problem.
Adds a simpler solution to the problem of mutable 'FileRule' state.

This will have to do for now. I just can't seem to reason about a better
solution that *actually* fixes the underlying problems that cause this.
This will have to be revisited at a later stage.

Simply do a deep copy of the 'Configuration' instance containing the
'FileRule' instances whose scores must not be shared between files.

Will likely need to rewrite this to work well with many of the upcoming
changes. But doing so at that time, when it is easier to think about
possible requirements, probably makes more sense than struggling now.