Fix recursion with unset or empty input paths bug.
Fixes a bug where empty or unset input paths would expand.

If the input paths is "" or unset, the 'normpath' function
would expand the empty path to the current working directory.

When running with recursion enable, thousands of files was
processed even though the user did not specify a single file.

This commit fixes this bug.