Remove the 'MetadataSourceConfigFieldParser' class.
Removes the 'MetadataSourceConfigFieldParser' class.

This field parser class does not follow the intended abstraction
model where the parser classes are mapped to expected data types
based on the "query string". The "metadata fields" could contain
any type of data, and so this class does not work well with the
rest of the current implementation.

Refers to TODO-list item [TD0048].