Pass extractor results data through custom types.
Adds passing extractor results data through the 'Path' type.

Also adds modifications to the 'FileObject' class to match these
changes. The 'FileObject' fields that are used in the changed
lines should already be in the "internal bytestring" format, and
so passing the data through the 'Path' type class returns them as-is.

However, the idea is that this is part of a overarching change in how
types are handled throughout the program and this specific case might
be revisited and simplified at some later stage.

Refers to [TD0044][TD0004].