Remove method 'get' from class 'AnalysisResults'.
Removes redundant method 'get' from the 'AnalysisResults' class.

Adds changes to use the 'query' method instead, that already
performed essentially the same task of providing data to outsiders.

Also removes raising exceptions when trying to fetch data using an
"invalid" label (query string) is as this was handled in the 'get'
method but not in the 'query' method.

References [TD0021].