Disable the 'CommonFileSystemExtractor' extractor.
Disables the 'CommonFileSystemExtractor' extractor temporarily.

This class is implemented to accept a 'FileObject' as the "source"
but the superclass and all other extractors expect a file path in
the "internal bytestring" format.

And so this extractor class raises an exception due to the unexpected
type of "source". This must be fixed by a major overhaul or simply
reverting to the previous method of collecting 'FileObject' data in
the 'Extraction' instance, before executing the extractor run queue.

Setting the MIME type globs to 'NONE/NONE' effectively disables this
extractor for now.

References [TD0051].