Substitute bad recursive function.
Substitutes a bad recursive function with a new common utility.

Both the 'AnalysisResults' and 'ExtractedData' classes used a
recursive function to count the number of contained items.
One of these implementations was bad and returned bad counts.

This commit adds a single shared utility function used by both
classes, as well as unit tests of this new function.