Add new type wrapper class 'PathComponent'.
Adds a new custom type wrapper class 'PathComponent'.

Passing parts of file names that could be empty, like a file extension
that might be missing without considering the file name to be invalid,
would trigger exceptions in the 'Path' type class that expected
full absolute paths.

This commit adds a new type class 'PathComponent' that accepts parts
of paths and even empty paths, like that of a missing file extension.