Disable logging in 'plugins' to restore logger.
Disables all use of logging in 'plugins/*.py'.

Disabling all use of logging in 'plugins/__init__.py' and
'microsoft_vision.py' solves the weird problem where the log format is
garbled and contains ':root:'.
Also, the command-line options that change the log format
('--verbose', '--debug') are ignored when the problem is in effect.

Removing all use of logging in 'microsoft_vision.py', where the log
format IS modified if the script is started stand-alone, does not solve
the problem.

The problem only goes away when all use of logging is disabled in
both 'plugins/__init__.py' and 'microsoft_vision.py'.