Improve formatting '__str__' of 'DataContainerBase'.
Adds nested format of 'DataContainerBase' (+subclasses) '__str__'.

Adds a new utility function 'expand_query_string_data_dict'.
This function performs the reverse operation of 'flatten_dict'.
I.E. it takes a dictionary on this form:

    input_dict = {
        'foo.bar.a': 1,
        'foo.bar.b': 2,
        'foo.baz.a': 3
        'mjao': 4
    }

And expands (deflates, "unflattens") it to the following:

    output_dict = {
        'foo': {
            'bar': {
                'a': 1,
                'b': 2,
            }
            'baz': {
                'a': 3
            }
        }
        'mjao': 4
    }

This functions is used in the '__str__' method of the
'DataContainerBase' base class, which in turn is inherited from by both
the 'AnalysisResults' and 'ExtractedData' classes.

Listing results now uses the string-representations of these container
classes instead of using the YAML module through calling 'util.dump' ..