Swallow all output in evaluation functions.
Adds subshell-wrap and redirecting stdout and stderr to /dev/null.

Adds wrapping the "eval" statement in a subshell, along with excessive
redirection intended to handle some tricky cases (try redirecting the
output of the "time" command, for instance) as well as an additional
redirection of the subshell output to /dev/null.

I'm pretty sure this isn't completely right, but test results look OK.