Modify the 'add' method in 'ExtractedData'.
Adds warning when passing empty data, fix None-test and refactoring.

Part of finding how to best test for missing data passed to the
'collect_results' callback method. Could test in either the callback
or in the 'ExtractedData' container class.

Fixes the test for empty data to explicitly test for "is None" so
that valid but "falsy" data like 'False' is passed through as-is.