Modify the 'BasePlugin' class.
Adds (WIP) modifications of the 'BasePlugin' class.

Removes having to pass 'source' to '__init__' so that plugins can be
instantiated prior to passing any specific sources to act on.

Adds a method 'test_init' that should return True if the plugin was
initialized properly. This could fail if some dependency like an
executable is missing.

When the plugin-system is first started, all plugins should be
enumerated and placed into some sort of list for later lookup.
During this process, the 'test_init' methods should be used to test
the plugin before considering it to be "available".