Remove "adapter" function. Rename variable. Fixes.
Removes the function '_collect_file_data' and adds various fixes.

The '_collect_file_data' function was used to "fill in" the current
'FileObject' to be used as the storage "key". This is now done by the
users of the callback instead. They have access to the same 'FileObject'
so this "adapting" makes little sense.

* Renames 'add_pool_data_callback' to 'add_pool_data' for consistency.
* Minor fix to unit tests. Still broken though.

References [TD0074][TD0075].