Cleanup handling config encoding and exceptions.
Adds various fixes to handling of configuration exceptions.

Attempt at cleaning up inconsistencies in exception handling and
encoding of the configuration file paths. Move towards storing
the configuration paths as "internal bytestrings" everywhere and
use the 'syspath' and 'displayable_path' functions when the paths
are to be used or displayed.