Refer to "file rules" simply as "rules".
Simplifies the naming convention of configuration rules.

The concept of a 'FileRule' does not seem to make much sense,
there are only 'Rules'; Rename all (most) mention of "file.*rule"
in both the sources, comments and documentation.

This commit modifies the configuration syntax so that older
configuration files will *FAIL TO LOAD* if they are not manually
edited.  Simply replace 'FILE_RULES' with 'RULES' and it should
load properly.