Fix logging loosing track of the root logger.
Fixes errors where the 'logging' module would ignore settings.

The 'logging' module would sometimes forget settings and ignore the
set verbosity level. I suspected this was related to the explicit
manipulation of the Python PATH when importing the 'extractors',
'analyzers' and 'plugins' packages.

This commit fixes the logging reset by using a module-level logger in
modules that are distinctively separated from the "core".