Handle exceptions better in 'NameBuilder'.
Improves exception handling in the 'NameBuilder' class.

the 'populate_name_template' function raises a 'NameTemplateSyntaxError'
if the name template can not be populated and returned because this
function is used by the field parsers to validate name templates in
the configuration as it it loaded.

The 'NameBuilder' should raise a 'NameBuilderError' if it fails.
This commit adds catching the first exception and re-raising the
exception expected by the users of 'NameBuilder'.