Fix crashing when repeating the same input path.
Fixes crashing when the same input path is given multiple times.

After normalizing the incoming path arguments to full absolute paths,
they are deduplicated by storing in a set before converting back to
the previously used list.

This solves the problem for now, but this kind of deduplication will
have to be reworked once again when adding safe handling of symlinks.