Change signature of 'NameBuilder' init.
Modifies the '__init__' method of the 'NameBuilder' class.

Instead of passing an active rule and using the name template and
data sources contained within, explicitly pass the name template
and data sources.

This means that simple dictionaries (data sources) and strings (name
template) are passed across boundaries, instead of the previous
passing of a 'Rule' class instance.