Do not keep state in 'Rule' instances.
Removes score and weight from the 'Rule' class.

Modifies the 'Rule' class and its users to not keep any state in the
'Rule' instances. These used to keep track of scores a 'score' property,
mutated through the 'upvote()' and 'downvote()' methods, as well as the
unimplemented 'weight' property.

These are now removed and the 'RuleMatcher' has been modified to store
scores and weights locally in a dictionary keyed by 'Rule' instances.

This hopefully solves having to to the following in 'rulematcher.py';

    self._rules = copy.deepcopy(active_config.rules)
    # TODO: Double-check that this isn't needed anymore, then remove.