Add resolving symlinks on MacOS.
Fixes running the wrapper script 'run.sh' as a symlink on MacOS.

The wrapper script would fail if one created a symlink to it and
ran the script from the script.

This commit fixes this by using Python to untangle any symlinks
and get an absolute path to the main autonameow directory.