Use type wrappers to format wrapped values.
Adds formatting 'ExtractedData' values in the name builder.

Adds using wrappers included with the 'ExtractedData' instances
to format the "raw" values. This solves converting MIME types
like 'image/jpeg' to 'jpg', suitable for use as an extension.

However, this does not solve the problem of converting MIME-types
to suitable file extensions. Multiple file extensions are mapped
to 'text/plain' for instance, so looking up 'text/plain' could
return the extension 'in' just as well as 'txt'.

References [TD0017].