Refactor the name builder. Add 'Resolver' class.
Factors out data collection from 'NameBuilder' to new class.

Adds a new class 'Resolver' in 'evaluate/resolver.py'.

Previously, the name builder was given a name template and a dict
mapping template fields to "MeowURIs". The name builder used this
to query the repository with the "MeowURIs" to collect data to be
populated in the name template.

This commit factors the data collection task out into a new
'Resolver' class. The idea is to separate collection from name
building, the former often fails and the upcoming interactive mode
will require a lot of extra functionality to be added, which would
not be possibly without some kind of refactoring, like this one.