Remove state and modify calling extractor classes.
Modifies extractors to not keep any state and simplifies use.

Previously, the extractor classes was handed a data source at
instantiation through the '__init__' method.
This meant that the classes first had to be "loaded" with a source
before they were executed in a later step, when running the "queue".

This commit modifies all extractor classes to not require any
arguments to the '__init__' method.
Instead, the data source is passed to the class 'start'-method.
This is done with the '__call__' method.

Old way of extracting data:

    extractor = DataExtractor(source)
    result = extractor.execute()

New way of extracting data:

    extractor = DataExtractor()
    result = extractor(source)

This commit also removes the "extractor run queue" as it no longer
required with this new way of executing extractors.

Extractors also stored data in instance variables such as 'self._data'.
This is also removed -- extractors should be called as functions.
Extractors should not keep any state between calls.