Remove keyword argument 'field' from 'execute'.
Removes keyword argument 'field' from extractors 'execute' method.

This optional keyword argument is never used.
It was originally intended to serve as a sort of caching mechanism;
the extractor would do an initial extraction and store the results
in a private attribute of the class.
However, since the extractors have been rework to not keep any
state, this type of caching is no longer viable.

The 'field' argument was also thought as a possible means of
optimization; a caller could ask for a subset of all possible
extracted data for performance reasons. In practice, this either
doesn't work out or isn't worth doing.

The 'field' argument is now removed, any optional arguments could
still be passed through to the extractors with '**kwargs'.