Add partial fix for bug where 'wrapper' is None.
Adds a partial fix for crashes when 'wrapper.formatter' is None.

If a extractor class does not provide any type wrappers when
constructing the 'ExtractedData' instances, and the auto-detection
of wrapper from the "raw" data type ALSO fails, the 'ExtractedData'
instance attribute 'wrapper' will be None.
Calling any method on the 'wrapper', like the name builder does
with 'wrapper.format'; will then result in a crash ..

This commit adds a partial fix and TODO-notes referencing [TD0088].