Modify text extractors to return dict.
Modifies text extractors 'meowuri_root' and common return method.

The repository 'store' method (and '_store' method) handles too many
specific cases, this is part of an effort to have non-core components
use a more consistent format for the data to be stored.

Instead of having the 'meowuri_root' be 'contents.textual.raw_text' and
returning the data "as-is" from the 'AbstractTextExtractor.excute()'
method, this commit modifies the 'meowuri_root' of all text extractors
to 'contents.textual' ('contents.visual' for the OCR-extractor..) and
adds returning a dict from the 'AbstractTextExtractor.excute()', with
'raw_text' as key.