Move "flattening" data out of 'Repository.store()'.
Removes "flattening" of in the 'Repository' before being stored.

Moves this to the local 'collect_results' functions that is passed
to extractors and analyzers.

This is motivated by attempting to minimize uncertainty of which
types of data structures the 'Repository' receives.
The "non-core" components have been changed so much, with so many
hack fixes and partial experiment added on, so that a common
interface barely exists any longer. This must be improved.