Return None from 'Configuration.get()'.
Replaces raising 'KeyError' with returning None for bad accesses.

Fixes bug causing program crashing when the name builder calls
'config.get(['CUSTOM_POST_PROCESSING', 'replacements'])' and the
configuration file does not contain these sections.

The configuration usually substitutes default values to prevent
this, but in this case it really should be up to the user to decide
if any substitutions should be made.