Fix major bug in 'ExtractedData'.
Fixes serious bug in 'ExtractedData'. Fix namespace.

The 'ExtractedData' instances would be shared when called
multiple times from the same "wrapper_tag_lookup" dictionary.

This commit adds a hack fix class method that uses 'deepcopy'
to make a copy of the instance before calling and returning it.

The type "wrappers" are renamed to "coercers" to avoid namespace
issues with instances of 'ExtractedData', that also are referred
to as "wrappers".