Rework handling "MeowURIs".
Modifies overall handling of "MeowURIs".

Instead of hard-coding a "MEOWURI_ROOT" in the source classes,
this new method uses the Python module hierarchy to construct
default MeowURIs.

It is still possible for source classes to override either parts
or the entire MeowURI.

The new convention is to always include the type of sources in
"source-specific" MeowURIs.

The first part is either 'extractor', 'plugin' or 'anlyzer'.

Example:  'extractor.metadata.exiftool.EXIF:DateTimeOriginal'

This makes a clear distinction between the specific MeowURIs and the
"generic" URIs, which will be a lot simpler; 'metadata.DateCreated'