Fix broken unit tests after 'Repository' changes.
Fixes a bug introduced in commit ca4779cdb7e6235911c799cf3081fa80669b.

Unit tests did not initialize the 'Repository'.

This commit fixes this by adding an additional 'meowuri_class_map',
local to the function 'map_meowuri_to_source_class'.
There really was no reason for the previous version to reach in to the
'Repository' class to access the objects attribute with the same name.