Rename 'ImageOCRTextExtractor' class and sources.
Renames the 'ImageOCRTextExtractor' to 'TesseractOCRTextExtractor'.

Renames the source file from 'ocr.py' to 'tesseractocr.py'.

Part of moving towards the new scheme of devising MeowURIs from
the module/package names in the hierarchy.

The new names distinguish this exact extractor, as opposed to the
old names that was intended to work with any OCR back-end.