Add new extractor 'JpeginfoMetadataExtractor'.
Adds a new extractor class 'JpeginfoMetadataExtractor'.

Returns a new type of generic field; 'health', that represents
the integrity or 1/corruption of a file as a float in range 0-1.

It also returns a related boolean result 'is_jpeg' that could be
used to verify file types, ensure quality and detect corrupt files.
Which is very useful when dealing with jpegs ..