Pass extractors 'FileObject', not bytestrings.
Replaces bytestring extractor sources with instances of 'FileObject'.

Extractors have been passed sources as absolute bytestring paths.
The 'FileObject' instance has the equivalent absolute path stored as an
attribute, and hopefully provides additional advantages.

This is motivated by replacing identifiers used by the caching system as
keys into the cached data. When the extractors were passed only an
absolute path to the source, they would have to make an extra call to
get some kind of checksum from the path, to use as the cache key.

By passing instances of 'FileObject', a reused method for getting a file
hash can be utilized. This is probably the '__hash__' method or a
separate hashing method in the 'FileObject' class.