Remove 'deepcopy' of rules in 'RuleMatcher'.
Removes using 'deepcopy' to get a copy of the rules.

This was necessary when the 'Rule' class kept scores and other state
that would be modified between 'RuleMatcher' evaluations of different
files. This is no longer the case, so this should not be needed and
does in fact not seem to be needed after a cursory glance.