Modify '__hash__' in the 'FileObject' class.
Modifies hashing 'FileObject' instances.

Adds using the "partial" cryptographic hash when hashing instances
of 'FileObject'.

When combined with the other attributes used in computing the hash,
this addition greatly reduces the risk for unwanted hash collisions
in practical usage.

Although, this should probably still be extended with the WAY WAAY
SLOWER method of hashing the entire file for critical applications.