Add changing persistence options in the config.
Adds a new section to the configuration file.

This new section allow specifying paths for the caching system
as well as the interactive prompt history.

    PERSISTENCE:
        cache_directory: null
        history_directory: null

If omitted or left empty, default values will be substituted.

NOTE: Default values are always '/tmp/autonameow_cache' and
      should be made platform-independent at some point.

This commit completes TODO-list item [TD0097].