Refactor main program entry point.
Separates the CLI-specific main entry from the "real" main entry.

Adds a new function 'real_main' that will be the _actual_ program
entry point.

This function is intended to be platform/interface-agnostic, in
that it should be called by an outer interface-specific layer; like
the CLI entry function 'cli_main'.
The passed in options are also in an internal format (dict) as to
not depend on any means of providing program options; argument
parsers, etc.
Default options are now defined in 'real_main' and checking option
combinations is also handled here.

This commit also separates the 'argparse' CLI argument parsing from
initializing the logging system.