Disable the 'PyPDFMetadataExtractor'.
PyPDF somehow overrides default warning systems. Disable for now.

It seems very difficult to disable warning messages emitted by PyPDF2
due to the way the library is designed. This, combined with the overall
poor performance and the fact that the project is currently
"unmaintained in the short term", this extractor is now disabled.