Move contents out of 'core/ui/cli/__init__.py'.
Adds new file 'core/ui/cli/common.py'.

Moves contents of 'core/ui/cli/__init__.py' to 'core/ui/cli/common.py'.

This package is intended to be the interface to multiple containing
sub-packages that each implement different UIs.

There is currently only a command-line interface, so this interface
does not exist. BUT it might be good to implement the interface only
once another UI is added and the real use and requirements are better
known anyway.