Fix crash. Add 'multivalued' to 'ExtractedData'.
Fixes crashing (sanity check failure) when formatting tags.

The name template field 'tags' is assumed to be a list, enforced by a
sanity check assertion. This failed when the filename analyzer passed
a single tag to the results.

This is fixed by modifying the 'ExtractedData' class to allow
coercing "multivalued" data. The relevant methods in the name builder
and template field classes are also modified to handle multiple values
differently; instead of passing a list of 'ExtractedData' instances,
a single instance of 'ExtractedData' now can contain a list of data.