Remove text extractor 'PyPDFTextExtractor'.
Removes the 'PyPDFTextExtractor' extractor.

This extractor has been disabled for some time due to poor results.
It is now removed. Should probably be replaced with some other PDF
text extraction method to provide an alternative to 'pdftotext' for
cases where it is not available.

But 'pdftotext' will have to do for now.