Add partial separation of CLI-specifics from UI.
Adds rework of module/package structure and imports.

* Replaces calls to 'cli.msg()' with 'ui.msg()'.
* Moves 'core/options.py' to 'core/ui/cli/options.py'.

The 'ui' package is intended to provide an interface to any UI.
Even if this is currently not relevant and future addition of an
alternative to the only current command-line interface UI, would
require reworking a lot of stuff due to incorrect assumptions;
enforcing this separation at this stage improves readability, etc.

Note that code inside 'ui/cli' that needs to call 'msg()' does not
call through the interface with 'ui.msg()', but instead uses the
CLI-specific function 'cli.msg()' directly.

Any other code that is outside of 'ui/cli' should call 'ui.msg()'.