Update regression runner and utils.
Adds various fixes to the regression test results printing.

Also adds skipping test by creating a file called 'skip'
in the relevant regression test directory. The file contents does
not matter, the test if skipped if the file exists.
This makes it pretty easy to disable a test with 'touch skip'.