Fix multiple sources was overwritten.
Adds temporary fix to bug where multiple MeowURIs was clobbered.

Multiple sources for a name template field in the configuration
file would not be properly parsed in 'config/rules.py'.
Only the last source would be used, clobbering the previous.

This commit fixes this by storing the MeowURI sources in a list,
which also required changes in the "resolver". These changes is
temporary in that the resolver is a terrible mess, and must be
cleaned up and/or rewritten in conjunction with reworking usage
of the 'ExtractedData' class.