Remove passing analyzer results through callbacks.
Removes passing analyzer results back through a callback.

Replaces the callback-method with returning results stored in an
instance attribute dict 'results'.

The 'run()' method is reworked to call a new method 'analyze()' that
does what 'run()' did previously, and then return the results.

This changes is motivated by making it easier to follow what is
going on and simplify the overall workings of the results collection,
that is currently undergoing various refactorings.

Completes TODO-list item [TD0122].