Modify 'unit_runner.sh' to use 'unittest'.
Replaces 'pytest' with 'unittest' in the unit test runner script.

Using 'pytest' was originally only due to the automatic generation
of HTML test reports for the course work. Now, there is no need for
the 'pytest'-specific functionality.

The reason for switching is that 'pytest' did not run a couple of
hundred of the unit tests. I'm not sure why but I noticed that the
PyCharm test runner ran a lot more tests than what 'unit_runner.sh'
did.