Rework format of returned extractor data.
Modifies extractors to return only a dict of values.

Instead of returning a dict with both values and additional
meta-information, the extractors now only return a dict keyed
by "fields" or ("MeowURI leafs") storing coerced data.

The additional meta-information stored in the class attribute
'FIELD_LOOKUP' is added to this separately in the calling code
'core/extraction.py'.