Fix bug in name template field 'Tags.format()'.
Fixes bug in the format method of the 'Tags' template field.

The format-method worked fine when passed a list of strings, but
when passed a single string, the string was split up and joined
with the "between-tag-separator", causing "tag" to become "t a g",
assuming a single space as "between-tag-separator".

This commit fixes this by coercing the data to a list of strings.

Added regression tests '0003_filetags_b' and '0003_filetags_c'
that demonstrates this failure case.