Modify regression runner. Add '--stderr' option.
Modifies the regression runner and adds a new option.

Moves catching the "top-level" exception to the wrapper class so
that the captured stdout and stderr are stored even when the main
program class raises an exception.

Adds a new option '--stderr' that prints the captured std error
output after running each test.