Fix 'ISBNMetadata' deduplication in 'EbookAnalyzer'.
Fixes instances of 'ISBNMetadata' not de-duplicated properly.

The comparison method of the 'ISBNMetadata' class was not used
when storing instances in a set. This commit fixes this by storing
the instances in a list after explicitly testing if the instance
has already been added, which actually uses the comparison method.