Add major refactoring of several parts.
Adds changes to the rule matcher, main class and rules.

Reworks the rules and rule matcher to integrate properly with
the changes to the main 'Autonameow' class. Main change being how
the rule matcher is used.

Previously, it was re-instantiated for every file by passing the
rules and the current file object to the rule matcher init.
Now, a single instance is passed the rules from the configuration,
which will only be read from. The rule matcher is then started by
calling the 'match()' method with a file object to match against the
rules. This method now returns all matching rules as a list of
tuples, each containing a rule, the rule score and the rule weight.

This list is sorted by priority, so grabbing the first element is
equivalent to the old code that accessed the 'best_match' property.

The main 'Autonameow' class logic is reworked quite a bit, working
towards the behaviour described in 'notes/modes.md'.
NOTE: Changes to the 'Autonameow' class is still a work in progress!

This commit also contains other miscellaneous changes, along with
unit tests for the rule matcher.