Add keeping track of failed ISBN metadata requests.
Adds caching failed ISBN metadata requests.

Avoids hammering the servers with the same failing requests.

This assumes that the request failed due to an invalid ISBN number
and not because the internet connection is down. This must be fixed,
especially since the cached ISBN metadata is stored in the same file
as the blacklist ..