Move integration tests to separate directory.
Moves integration tests to new directory 'tests/integration'.

A lot of scripts are changed due to complications when handling sourcing
and paths of bash scripts. It is not really possible to get the absolute
path of the current script that works in all cases; running the script
from the same directory, using a relative path, absolute path, through
a symlink, etc..

This commit adds a simpler solution where one script exports global
environment variables that are used by the other scripts.
This is handled by the new 'setup_environment.sh'.

This new arrangement means that integration test suites no longer can
be executed as stand-alone scripts.
The complexity and workarounds just isn't worth the hassle.

Integration tests should be executed with 'integration_runner.sh'.